Language.C.Clang

Index

data ClangIndex

createIndex :: IO ClangIndex

Utilities

class Clang n

Minimal complete definition

deref, unsafeToPtr

newtype ClangOrd a

The Eq instance for Clang types checks structural equality, i.e. whether they represent the same object in the translation unit.

Wrapping values in this type provides Eq and Ord instances based on reference equality.

Constructors

ClangOrd

Fields

getOrdered :: a